requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
python-dotenv==1.0.0

# Web scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3
fake-useragent==1.4.0

# AI and NLP
openai==1.6.1
tiktoken==0.5.2

# Data processing
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# Validation and utilities
pydantic==2.5.2
email-validator==2.1.0
python-dateutil==2.8.2

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Database (optional for future)
SQLAlchemy==2.0.23

